Add ByteTracker and DeId classes, and create UI for video editing
This commit includes the implementation of ByteTracker and DeId classes for detection tracking and de-identification respectively. In addition, a comprehensive UI has been created for video editing which allows for video playback, analysis, de-identification, and result preview. Capability to open and close video files through the menu bar is also added.