Add README and LICENSE files
This commit includes the addition of the detailed README file for the VideoMasker project. It outlines details about the project, how to get started, and other relevant information for users and contributors. It also includes the addition of the GNU Affero General Public License to the repository, ensuring the protection and freedom to share and change the software.