Update tracking method and clean up unused code
The SFSORT tracking method was commented out and the BYTETracker falls back as the default method. Also, the code related to the EmbedDetector class was removed due to remove unused portion, and numpy with pillow were added to the requirements. The condition for using cuda has been slightly updated as well.